feat: split into separate dialogs
